requirements.txt
yt-dlp>=2024.0
pydub>=0.25.1
numpy>=1.24.0
faster-whisper>=1.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pywebview>=5.0
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
